refactor(sql): Allow multiple drivers at the same time
