+++ .changes/feat-multiple-sql-backends.md
@@ -1,0 +1,5 @@
+---
+sql: patch
+---
+
+It is now possible to enable multiple SQL backends at the same time. There will be no compile error anymore if no backends are enabled!
