+++ plugins/sql/Cargo.toml
@@ -30,6 +30,8 @@ tokio = { version = "1", features = ["sync"] }
 indexmap = { version = "2", features = ["serde"] }
 
 [features]
+#default = ["sqlite", "mysql", "postgres"]
 sqlite = ["sqlx/sqlite", "sqlx/runtime-tokio"]
 mysql = ["sqlx/mysql", "sqlx/runtime-tokio-rustls"]
 postgres = ["sqlx/postgres", "sqlx/runtime-tokio-rustls"]
+# TODO: bundled-cipher etc
