+++ plugins/sql/src/decode/mod.rs
@@ -3,17 +3,8 @@
 // SPDX-License-Identifier: MIT
 
 #[cfg(feature = "mysql")]
-mod mysql;
+pub(crate) mod mysql;
 #[cfg(feature = "postgres")]
-mod postgres;
+pub(crate) mod postgres;
 #[cfg(feature = "sqlite")]
-mod sqlite;
-
-#[cfg(feature = "mysql")]
-pub(crate) use mysql::to_json;
-
-#[cfg(feature = "postgres")]
-pub(crate) use postgres::to_json;
-
-#[cfg(feature = "sqlite")]
-pub(crate) use sqlite::to_json;
+pub(crate) mod sqlite;
